+++ src/views/header.php
@@ -6,5 +6,13 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <link rel="icon" type="image/png" href="./src/assets/images/icones8-favicon.png" />
-  <title>[S11] Selection BTS 0.1.0</title>
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" />
+  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Questrial&display=swap" />
+  <link rel="stylesheet" href="./src/assets/css/style.css" />
+  <title>[S11] Selection BTS 0.1.1</title>
 </head>
+
+<body>
+  <div class="hero is-success is-fullheight">
+    <div class="hero-body">
+      <div class="container has-text-centered">