+++ CHANGELOG.md
@@ -5,3 +5,21 @@
 ### Added
 
 - Application added to the repository
+
+## 1.0.0 (2021-12-28 to ...)
+
+### 0.1.1 (2021-12-28)
+
+#### Added
+
+- CSS stylesheet at `./src/assets/css/style.css`.
+- Common Footer file at `./src/views/footer.php`.
+
+#### Updated
+
+- `./src/views/login/loginHome.php` has complete CSS.
+- `./src/views/header.php` :
+  - CSS file;
+  - FontAwesome;
+  - Questrial Google Font;
+  - First common ```<div>``` tags;